Added interdiction torch for all entities Added configs for torches Added torches models and textures Added localisation for torches Added recipes for torches and minigun ammo Fixed recipe for bullets

+++ src/main/java/com/brightspark/bitsandbobs/reference/Config.java
@@ -37,5 +37,5 @@ public class Config
     public static int interdictionTorchRadius = 4;
 
     /** The strength for the interdiction torches */
-    public static float interdictionTorchStrength = 0.5F;
+    public static double interdictionTorchStrength = 0.5D;
 }
